python-examples/requirements.txt
# Core dependencies for WebSocket examples
websockets>=11.0.0
python-dotenv>=1.0.0